refactor of flow.py
