+++ src/t2r2/__init__.py
@@ -1,1 +1,1 @@
-from t2r2.flow import get_metrics, init, loop, dvc_checkout, dvc_metrics, dvc_params
+from t2r2.flow import dvc_checkout, dvc_metrics, dvc_params, get_metrics, init, loop
