+++ src/t2r2/config/__init__.py
@@ -1,1 +1,1 @@
-from t2r2.config.config import Config, load_config
+from t2r2.config.config import Config, load_config, get_config
