+++ src/t2r2/dataset/common.py
@@ -65,7 +65,11 @@ class WithMetrics:
         return flatten_dict(
             {
                 metric.name: get_metric(metric.name)(
-                    true_labels, predictions, **metric.args, proba_predictions=proba_predictions, stage=self.stage
+                    true_labels,
+                    predictions,
+                    **metric.args,
+                    # FIXME: this doesn't work without ignore_unmatched_kwargs
+                    # proba_predictions=proba_predictions, stage=self.stage
                 )
                 for metric in self.metrics
             }
